// weather-dashboard/src/types.ts
export type Place = {
  name: string;
  country_code: string;
  latitude: number;
  longitude: number;
};

export type WeatherResponse = {
  current: {
    temperature_2m: number;
    relative_humidity_2m: number;
    apparent_temperature: number;
    weather_code: number;
    wind_speed_10m: number;
  };
  daily: {
    time: string[];
    weather_code: number[];
    temperature_2m_max: number[];
    temperature_2m_min: number[];
  };
};